requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.6.0
pytz==2022.7
requests==2.28.1
twilio==7.16.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
openai==0.25.0



python-dotenv
